[syntax] implement intellisence specific TokenTypeExtension
