+++ VSRAD.Syntax/Core/Tokens/TokenType.cs
@@ -42,27 +42,6 @@ namespace VSRAD.Syntax.Core.Tokens
 
     public static class RadAsmTokenTypeExtension
     {
-        public static string GetName(this RadAsmTokenType type)
-        {
-            switch (type)
-            {
-                case RadAsmTokenType.FunctionParameter:
-                    return "argument";
-                case RadAsmTokenType.Label:
-                    return "label";
-                case RadAsmTokenType.GlobalVariable:
-                    return "global variable";
-                case RadAsmTokenType.LocalVariable:
-                    return "local variable";
-                case RadAsmTokenType.FunctionName:
-                    return "function";
-                case RadAsmTokenType.Instruction:
-                    return "instruction";
-                default:
-                    return "unknown";
-            }
-        }
-
         private static readonly Dictionary<RadAsmTokenType, string> _classificationTypeNames = new Dictionary<RadAsmTokenType, string>()
         {
             { RadAsmTokenType.Comment, PredefinedClassificationTypeNames.Comment },
